Add Customer

+++ app/Http/Controllers/CustomerController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\CustomerRequest;
 use App\Http\Traits\GeneralTrait;
 use App\Models\Customer;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class CustomerController extends Controller
 {
@@ -17,10 +18,7 @@ class CustomerController extends Controller
      */
     public function index()
     {
-        $customers = Customer::all();
-        return view('customer.index', [
-            'customers' => $customers,
-        ]);
+        return view('customer.index');
     }
 
     public function fetchCustomers(){
@@ -37,9 +35,6 @@ class CustomerController extends Controller
      */
     public function create()
     {
-        return view('customer.create', [
-            'customer' => new Customer(),
-        ]);
     }
 
     /**
@@ -48,11 +43,40 @@ class CustomerController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(CustomerRequest $request)
+    public function store(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+            'father_name' => 'required',
+            'type' => 'required',
+            'marital_status' => 'required',
+            'cell' => 'required',
+            'cnic' => 'required',
+            'monthly_income' => 'required',
+            'residential_address' => 'required',
+            'office_address' => 'required',
+            'caste' => 'required',
+            'cnic_expiry' => 'required',
+            'dob' => 'required',
+            'occupation' => 'required',
+            'designation' => 'required',
+            'work_since' => 'required',
+            'residential_phone' => 'required',
+            'residential_since' => 'required',
+            'office_phone' => 'required',
+            'cnic_front' => 'sometimes|file|image',
+            'cnic_back' => 'sometimes|file|image',
+            'image' => 'sometimes|file|image',
+        ]);
+        if (!$validator->passes()){
+            return response()->json(['status' => 0, 'error' => $validator->errors()->toArray()]);
+        }
+
         $customer = Customer::create($request->all());
         $this->storeImage($customer);
-        return redirect(route('customer.index'));
+        if ($customer){
+            return response()->json(['status' => 1, 'message' => 'Customer Added Successfully']);
+        }
     }
 
     /**
@@ -72,11 +96,25 @@ class CustomerController extends Controller
      * @param  \App\Models\Customer  $customer
      * @return \Illuminate\Http\Response
      */
-    public function edit(Customer $customer)
+    public function edit($customer)
     {
-        return view('customer.edit', [
-            'customer' => $customer,
-        ]);
+        $customer = Customer::find($customer);
+        $type = $customer->getAttributes()['type'];
+        $marital_status = $customer->getAttributes()['marital_status'];
+        if ($customer){
+            return response()->json([
+                'status' => 200,
+                'customer' => $customer,
+                'type' => $type,
+                'marital_status' => $marital_status,
+            ]);
+        }
+        else{
+            return response()->json([
+                'status' => 404,
+                'message' => 'Customer not found'
+            ]);
+        }
     }
 
     /**
@@ -86,8 +124,35 @@ class CustomerController extends Controller
      * @param  \App\Models\Customer  $customer
      * @return \Illuminate\Http\Response
      */
-    public function update(CustomerRequest $request, Customer $customer)
+    public function update(Request $request, $customer)
     {
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+            'father_name' => 'required',
+            'type' => 'required',
+            'marital_status' => 'required',
+            'cell' => 'required',
+            'cnic' => 'required',
+            'monthly_income' => 'required',
+            'residential_address' => 'required',
+            'office_address' => 'required',
+            'caste' => 'required',
+            'cnic_expiry' => 'required',
+            'dob' => 'required',
+            'occupation' => 'required',
+            'designation' => 'required',
+            'work_since' => 'required',
+            'residential_phone' => 'required',
+            'residential_since' => 'required',
+            'office_phone' => 'required',
+            'cnic_front' => 'sometimes|file|image',
+            'cnic_back' => 'sometimes|file|image',
+            'image' => 'sometimes|file|image',
+        ]);
+        if (!$validator->passes()){
+            return response()->json(['status' => 0, 'error' => $validator->errors()->toArray()]);
+        }
+        $customer = Customer::find($customer);
         $customer->update($request->all());
         $this->storeImage($customer);
         return redirect(route('customer.index'));
@@ -99,10 +164,20 @@ class CustomerController extends Controller
      * @param  \App\Models\Customer  $customer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Customer $customer)
+    public function destroy($customer)
     {
+        $customer = Customer::find($customer);
+        if (!$customer){
+            return response()->json([
+                'status' => 0,
+                'message' => 'Customer not exist'
+            ]);
+        }
         $customer->delete();
-        return redirect(route('customer.index'));
+        return response()->json([
+            'status' => 1,
+            'message' => 'Customer Deleted Successfully'
+        ]);
     }
 
     public function storeImage($customer)
